app/requirements.txt
django
django_filters
djangorestframework
djangorestframework-yaml
djangorestframework-xml
djangorestframework-csv
dry-rest-permissions

markdown
django-filter
django-admin-bootstrapped>=2.4.0
django-authtools>=1.2.0
django-tinymce
Pillow
image
easy-thumbnails>=2.2

django-debug-toolbar
django-environ==0.3.0
django-localflavor==1.1
django-nose==1.4.1
django-braces
django-crispy-forms

pytz
sqlparse
#mysqlclient
django-picklefield
django-constance
django-constance[database]

hl7apy

uwsgi
supervisor-stdout